+++ packages/falcon-front-kit/src/DynamicRoute/index.ts
@@ -1,1 +1,2 @@
 export * from './DynamicRoute';
+export * from './SwitchDynamicUrl';
